feat(files): track source blocks count, check crc32, RequestCompleteUpload handler

+++ src/main/scala/com/secretapp/backend/data/message/rpc/file/RequestCompleteUpload.scala
@@ -2,7 +2,7 @@ package com.secretapp.backend.data.message.rpc.file
 
 import com.secretapp.backend.data.message.rpc._
 
-case class RequestCompleteUpload(config: UploadConfig, blockCount: Int, crc32: Int) extends RpcRequestMessage
+case class RequestCompleteUpload(config: UploadConfig, blockCount: Int, crc32: Long) extends RpcRequestMessage
 
 object RequestCompleteUpload extends RpcRequestMessageObject {
   val requestType = 0x16
